change irvin

+++ app/src/main/java/com/fivesys/alphamanufacturas/fivesys/context/dao/overMethod/AuditoriaOver.kt
@@ -1,6 +1,7 @@
 package com.fivesys.alphamanufacturas.fivesys.context.dao.overMethod
 
 import android.os.Environment
+import android.util.Log
 import com.fivesys.alphamanufacturas.fivesys.context.dao.interfaces.AuditoriaImplementation
 import com.fivesys.alphamanufacturas.fivesys.entities.*
 import com.fivesys.alphamanufacturas.fivesys.helper.Util
@@ -62,6 +63,10 @@ class AuditoriaOver(private val realm: Realm) : AuditoriaImplementation {
         return realm.where(Detalle::class.java).equalTo("AuditoriaDetalleId", AuditoriaDetalleId).findFirst()
     }
 
+    override fun getDetalleByAuditoria(AuditoriaId: Int, Eliminado: Boolean): RealmResults<Detalle> {
+        return realm.where(Detalle::class.java).equalTo("AuditoriaId", AuditoriaId).equalTo("Eliminado", Eliminado).findAll()
+    }
+
     override fun getDetalleIdentity(): Int {
         val detalle = realm.where(Detalle::class.java).max("AuditoriaDetalleId")
         val result: Int
@@ -74,6 +79,9 @@ class AuditoriaOver(private val realm: Realm) : AuditoriaImplementation {
             val dd: Detalle? = realm.where(Detalle::class.java).equalTo("AuditoriaDetalleId", d.AuditoriaDetalleId).findFirst()
             if (dd != null) {
                 dd.AspectoObservado = d.AspectoObservado
+                dd.CategoriaId = d.CategoriaId
+                dd.ComponenteId = d.ComponenteId
+                dd.AuditoriaId = d.AuditoriaId
                 dd.Categoria = realm.copyToRealmOrUpdate(d.Categoria!!)
                 dd.Componente = realm.copyToRealmOrUpdate(d.Componente!!)
                 dd.Detalle = d.Detalle
@@ -84,6 +92,7 @@ class AuditoriaOver(private val realm: Realm) : AuditoriaImplementation {
                 dd.S4 = d.S4
                 dd.S5 = d.S5
                 dd.Url = d.Url
+                dd.estado = d.estado
             } else {
                 val a: AuditoriaByOne? = realm.where(AuditoriaByOne::class.java).equalTo("AuditoriaId", AuditoriaId).findFirst()
                 if (a != null) {
@@ -96,18 +105,24 @@ class AuditoriaOver(private val realm: Realm) : AuditoriaImplementation {
 
     override fun deleteDetalle(d: Detalle): Boolean {
         var valor = true
-        val imagepath = Environment.getExternalStorageDirectory().toString() + "/" + Util.FolderImg + "/" + d.Url
-        val f = File(imagepath)
-        if (f.exists()) {
-            if (f.delete()) {
-                realm.executeTransaction {
-                    d.deleteFromRealm()
+        if (d.estado == 1) {
+            val imagepath = Environment.getExternalStorageDirectory().toString() + "/" + Util.FolderImg + "/" + d.Url
+            val f = File(imagepath)
+            if (f.exists()) {
+                if (f.delete()) {
+                    realm.executeTransaction {
+                        d.deleteFromRealm()
+                    }
+                } else {
+                    valor = false
                 }
-            }else{
+            } else {
                 valor = false
             }
-        }else{
-            valor = false
+        } else {
+            realm.executeTransaction {
+                d.Eliminado = true
+            }
         }
         return valor
     }
